[FIX] knowledge: make sidebar resizer compatible with rtl languages
Purpose:
========
Currently, the sidebar resizer is not compatible with rtl languages.
When a rtl language is used, the sidebar is placed on the right (while it
is placed on the left with ltr languages).
However, the width of the sidebar is set as the horizontal position of the
cursor when the resizer is moved (which is independent of the language
direction).

This commit changes the computation of the width of the sidebar when the
resizer is moved and when a rtl language is used.

Task-3131218

closes odoo/enterprise#36186

X-original-commit: fba7629931cc199b3eb84d03506c3bb2a4b4d4d4
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>