[FIX] knowledge: paste template content in the last blank line
Currently, when using the feature "use as description" from a template block in
knowledge, if the record's field to be filled is "empty" (contains an empty
paragraph), the content would be pasted after it.

With this fix, the content will replace that empty paragraph instead.

Furthermore, if the last line of the html_field has explicit `<br`> nodes (i.e.
the result of a SHIFT + ENTER after a text node:
```<p>content
<br><br></p>```

The pasted content will be positioned directly under the previous content
(discarding the blank `<br>`):
```<p>content</p>
<p>pasted</p>
<p><br></p>```

Task-3101553

Part-of: odoo/enterprise#34977