[FIX] account_accountant: Fix auto-balance line multi-currencies in bank reco widget
Steps to reproduce:
- Create a journal with a foreign currency
- Create a statement line on this journal using a rate
- Change the rate on the foreign currency to something else
- Open the bank reconciliation widget
=> The computed amount for the auto_balance line is wrong.

Currently, to generate the auto-balance line, we call the method used when creating the journal entry for the statement line.
This method is computing the company amounts using the rate encoded in odoo.
However, the liquidity line is using the journal entry linked to the statement line to compute its amounts.
To fix this inconsistency, we retrieve the transaction amount, journal amount and the company amount only from the accounting data without using any conversion rate.

closes odoo/enterprise#35723

Related: odoo/odoo#109782
Signed-off-by: Quentin De Paoli <qdp@odoo.com>