[FIX] account_accountant: prevent restore st line if record is unavailable
The bank reconciliation widget tries to restore the last statement line that was active when returning to the kanban. It's possible that the statement line is not available in the list of records, causing a stacktrace.
To replicate:
Accounting > Dashboard > Reconcile 8 Items
Switch to list view
Search for a transaction that returns 0 records for eg. yabadabadoo
Switch to kanban

With this fix, no record is selected and the no_content_helper screen is displayed instead

closes odoo/enterprise#35750

X-original-commit: 35cb7823a804d7ea500ef1088ebddcbbcf68da72
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Ayob Habib (ayh) <ayh@odoo.com>