[FIX] knowledge: allow typing when selecting an article to link
Currently, when using the "/article" command, you cannot type into the
"select2" custom element to filter on a specific article.

This is caused by a conflict with Bootstrap's "focustrap" that moves the focus
outside of the input, and does not allow to move it back in it.

This commit hacks the modal to disable this focustrap when the modal is shown.

Additionally, we force the focus on the search input so that the user may start
typing directly.

This will be refactored in master to avoid using a legacy modal.

Task-3079001

closes odoo/enterprise#35733

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>