[IMP] l10n_us_1099: make name field required in l10n_us.1099_box model
before this commit, the name field is not required in the model level/view level and empty records can be created in this table without filling any data.

after this commit, the name field will be required at the model level and without filling the name (the single field in the model), records cannot be created.

closes odoo/enterprise#35748

Signed-off-by: John Laterre (jol) <jol@odoo.com>