[FIX] project_sale_subscription: put subs services only in subs section
Current behaviour:
When selling a recurrent service with fixed/prepaid invoicing, it
appears two times in the project profitability report, once as
"Fixed/Prepaid" Service, another time as a "Subscription".

Expected behaviour:
Such service should only be present in the "Subscription" section.

Steps to reproduce:
- Install Sales, Project, Subscription
- Create a Service that is invoiced "Fixed/Prepaid", creates a new
  project on creation
- Create a new SO with that product, confirm it.
- Go to the Project's Update of the newly created project
- Observe that the services was accounted 2 times, onces in the
  "Fixed/Prepaid" section, another time in the "Subscription".

Reason for the problem:
When calculating the profitability from "Fixed/Prepaid" services,
we ignore the fact if they are a subscription or not.

Fix:
Extend the domain to not take subscription into account when
generating the profitability items that are prepaid services.

Affected versions:
- 16.0
- master

opw-3113320

closes odoo/enterprise#35765

Related: odoo/odoo#109862
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>