[FIX] web_studio: Allow users to use rule condition on approvals
Steps:
  - Go to CRM for example and select any record
  - Open studio in the form view
  - Add an approval rule on the `won` button with `admin` user
  - Close studio
  - Try to click on the `won` button with `demo` user
  - Approval is working
  - Try now with `admin` user to add a rule condition (for example `('display_name', 'is', 'Private')`)
  - Try again with `demo`
  - Approval is not working since the rule condition addition, even if the `display_name` is `Private`

Before this fix, approvals were updated like this

```js
patch(ViewButton.prototype, "web_studio.ViewButtonApproval", {
	setup() {
		this._super(...arguments);
		if (this.props.studioApproval) {
			//...
			useApproval({})
			//...
		}
	}
})
```

But overriding setup method is not enough since it will not be updated if we change records via the
pager buttons, nor in the case of a new record.

Using `owl.onWIllUpdateProps` we can detect the change of record ids and update the props in order
to always be on the right record`.

opw-3098478

closes odoo/enterprise#35760

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>