[FIX] knowledge: fix portal search bar
Purpose
=======
Fix the knowledge portal search bar which
is currently not finding any of the children
articles if one of its parents is folded.

Specification
=============
Renders a completely flattened tree with the
articles matching the search. A flattened tree
prevents noise (unnecessary parents display when
children are matching the search) and redondancy
(duplicated articles because of the favorite tree).

Task-2984564

closes odoo/enterprise#33811

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>