[FIX] planning: add new step to open scale dropdown in tour
Before this commit, the tour tries to click on the Day scale button but
the scales buttons have been reviewed and converted into a dropdown as
it is the case in the calendar view.

This commit fixes the tour by adding a step to click on the button to
open the scale dropdown.

Part-of: odoo/enterprise#35769