[FIX] planning: create shift the next day to be able to ask switch
Before this commit, the test in which the JS tour is started to test the
"Ask switch shift" feature, creates a shift in the current date with
start_datetime and end_datetime set between 23:50 and 23:55, and so the
tour JS could fail if it is started after 23:55 because the shift will
be in the past and so the button to ask switch shift won't be visible.

This commit fixes the issue by creating the shift the next day to be
sure the shift created will no longer be in the past when the tour is
launched.

closes odoo/enterprise#35769

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>