[IMP] knowledge: transition from select2 to SelectMenu
In this commit, we are transitioning the select2 inputs to an OWL version since
select2 is bound to be deprecated in the future.

In order to do that we are using a Component called SelectMenu. This Component
asks for a set of values linked to labels and it will search inside these
values when you input text inside the searchBar.

We needed to modify two dialogs that were using the old select2:
`MoveArticleDialog` and `ArticleSelectionBehaviorDialog`.

For `MoveArticleDialog`, we needed to use groups since this dialog also enabled
the user to move an article inside a specific section.
So we made two groups: an article group and a section group.
For the section group we added the two sections 'Private' and 'Workspace',
for the 'Private' section we added the image of the logged user beside its
label using the `choice` slot of the `SelectMenu`, this slot enables us to
customize each choice inside the `SelectMenu`.

For `ArticleSelectionBehaviorDialog`, we only used the articles as the choices
for the menu.
We also added a `Create` button to the menu using the new `bottomArea` slot
added by the JS team to the Component. This button creates a new article under
the one we are currently, or inside the workspace if the command is called from
outside Knowledge.
If we are inside Knowledge, we trigger a render of the article tree to display
the newly created article inside it.

For each dialog, we are using the default slot to display the selected article.

task-3134491

closes odoo/enterprise#38198

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>