[FIX] planning: fix JS tour launched in python test
Before this commit, the test, in which the
`planning_shift_switching_backend` tour is launched, a slot is generated
the next day from the commit [1]. By doing that, if the tour is started
on Saturday, the tour will fail because slot will no longer be in the
current week but in the following week in the gantt since, the last day
in a week in the gantt view is Saturday.

This commit creates the slot in the current datetime for a period set to
one hour to be sure the slot is not considered in the past during the
execution of the tour.

[1] b9e4be033f0a43c9c94d9efc898787bbadd57d12

closes odoo/enterprise#35782

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>