[FIX] account_reports : aged reports, amount currency
In the aged reports (aged receivable/payables), the column amount currency was present which might lead to misunderstanding for users. Indeed, the user expects to see the residual amount instead of the original amount when a partial payment has been done and reconciled. This commit changes the data retrieved from the db but no fields as it is a stable version

closes odoo/enterprise#35854

Task: 3105257
Pr: 35726
X-original-commit: e5e168d596aaefcd842fc553890e474aff3322f2
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>