[FIX] planning: fix name compute when planning too many hours
The method `plan_hours` with default demo data cannot handle more than
~8000 hours[1]. If we input, for example, 8745 hours we get an error.
What's worse, the data is saved to the DB but the record cannot be
displayed.

Here we show a validation error when the user tries to input a too long
duration value.

A new test ensures the compute of name actually validates the `duration`
value

[1]: https://github.com/odoo/odoo/blob/6d04ee7b09fa79556e2ed3487fa769164093f468/addons/resource/models/resource.py#L714-L715

closes odoo/enterprise#35866

X-original-commit: f5a31a8dbe6c95af86ec59035088e0397b5c17c3
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>