[FIX] timesheet_grid: prevent old timesheet start
This is a forward port of https://github.com/odoo/enterprise/pull/35186
(without the check on `prevent_old_thimesheets_encoding` as this setting
has been removed in https://github.com/odoo/enterprise/pull/34043)

It is possible to start a timesheet older than the employee's last
validated timesheet but an error is raised when trying to stop it

Steps to reproduce:
1. Install Timesheets
2. Go to Settings > Timesheets > Time Encoding and enable 'Lock Dates'
3. Set Marc Demo's rights on Project to 'User' and Timesheets to 'User:
own timesheets only'
4. Go to Projects and open any task
5. Add 2 timesheets (with a duration) for Marc Demo on different days
6. Go to Timesheets > To Validate > All Timesheets and validate the most
recent timesheet of Marc Demo (in the list view)
7. Connect as Marc Demo and open Timesheets
8. Trigger the list view
9. Start the older timesheet and then stop it: an error is raised

Solution:
When we try to start the timer of an old timesheet (older than the
`last_validated_timesheet_date` of the employee), create an AAL on
today's date for the same project/task. When validating timesheets, we
interrupt all the old timesheets.

opw-3095155

closes odoo/enterprise#35787

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>