[FIX] data_merge: fix merge bug for duplicated records
Currently duplicated records cannot be merged. This is because the
new _search_company_id method does not account for data_merge_records
whose related res_id has been deleted, which happens when merging
records.

To fix that, this commit add a condition to one of the subquery in
the form of a NOT EXISTS. The condition boils down to
 -> If res_id still exists in the database, check the res_id's
    company_id value.
 -> If res_id does not exsits, then treat it has if company_id = False.

Add a multi_company rule test to data_merge.

closes odoo/enterprise#35971

X-original-commit: 8b75dbe0ed6aaadba29c3f7a54156d029f31c73e
Signed-off-by: Kevin Baptiste <kba@odoo.com>
Signed-off-by: Van Delft Aurélien (avd) <avd@odoo.com>