[FIX] knowledge: fix kanban cards layout with long article names
Purpose:
========
Currently, when an article has a long name, the name overflows out of the
kanban card (both in the articles kanban view, and in the items embedded
kanban).
This commit prevents article names from overflowing their kanban cards
using the bootstrap text-truncate class.

This commit also prevents the name of embedded views of being hidden
behind the "rename" and "open" buttons, then reappear after these buttons
(with this commit, they are just hidden behind them, but one does not see
a letter of the embedded view name between these buttons and the search
bar).

Task-3138134

closes odoo/enterprise#35896

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>