[FIX] l10n_be_reports: proper handling of the VAT number when exporting EC Sales List with an accounting firm set
To reproduce:

1) Create a company in Belgium
2) Assign a VAT number to this company
3) Set an accounting firm as representative of this company. Make sure it has no VAT number
4) Open the EC Sales List and try exporting it to XML

==> An error message should be shown, warning that some data (the VAT) are missing on your representative, and proposing to complete them.

Instead, a traceback was raised. This was due to the fact _get_xml_export_representative_node received the EC Sales report as its parameter,
and tried to call _split_vat_number_and_country_code on its handler, while this function is only defined on the Belgian tax report's custom handler.

We solve that by revamping _split_vat_number_and_country_code, so that it can be called without referring to any handler.

OPW 3102753

closes odoo/enterprise#35944

X-original-commit: 130e3a203c4d78c534f10616fe84912211100d48
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>