[IMP] account_reports: Allow custom control of the automatic unfold behavior when printing a report
Since 16.0, printing a report to pdf or xlsx automatically unfolded all of its lines, leading to performance issues on large Aged Payable/Receivable (and other reports whose sublines result of a 'group by').

This behavior is intended on most reports, but we want to have more flexibility, and be able to change it case-by-case if necessary. This is done by slighlty modifying the way options['unfold_all'] is used, and the way options are handled when printing. We now always exclusively rely on this option key to know whether or not to automatically unfold, even when printing. Thanks to this change, we are able to modify options['unfold_all'] in a custom options initializer for any report, giving us the flexibility we want.

Aged Payable/Receivable now make use of this new feature to only unfold all their lines when printing if the 'Unfold All' filter was explicitly set by the user in the interface. Otherwise, by default, if no line is unfolded in the UI, none will be in the printed file either.

closes odoo/enterprise#36004

X-original-commit: 5787cd561028a8f5f0882c99316c429dfadfab10
Signed-off-by: Laurent Smet <las@odoo.com>