[IMP] knowledge: keep editor focus when clicking on "Add {icon,cover}"
When the user creates a new article, the editor will automatically be
focused. When the user clicks on the "Add icon" button or the "Add Cover"
button, the user will lose the editor focus which is annoying. This
commit will ensure that the user will not lose the editor focus when
clicking on those buttons.

task-3134527

closes odoo/enterprise#35788

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>