[FIX] industry_fsm: fix unexpected behavior due to refactor on views
Before this commit, the `To invoice` is applied by default in `All
tasks > To invoice` menu only when the user is in debug mode. This
behavior should be there even if the user is not in debug mode.

This commit fixes the filter in the search view to be able to
apply the filter by default when the user is in debug mode or not.

closes odoo/enterprise#36348

X-original-commit: 6365d573935fcadb620079e19cd3be4dcc3b9a0c
Related: odoo/odoo#111125
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>