[REF] even_barcode: convert legacy client actions to owl
The goal of this commit is to convert the last legacy client actions from
/even_barcode to owl.

Client action:
 even_barcode.event_barcode_scan_view

We take advantage of this commit to remove the addons 'event_barcode_mobile',
'hr_attendance_mobile' and 'barcodes_mobile' which is not necessary
anymore. The logic of this addon is necessary in desktop too. So we move it.

closes odoo/enterprise#35884

Taskid: 3138068
Related: odoo/odoo#110095
Related: odoo/upgrade#4296
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>