[FIX] account_avatax: don't check address on payments
`account.payment` inherits `account.move`. Its fiscal_position_id
field will be computed as normal. In the case of auto-applied fiscal
positions it's possible `account.payment` records will have Avatax
fiscal positions. When that happens a ValidationError is raised if the
partner's address is incomplete (missing country, state or zip).

To reproduce:

1/ set the Avatax fiscal position to auto-apply on California
2/ create a partner with state California and country USA but no zip
3/ create and validate an invoice for this customer
4/ click "Register payment" on the invoice
5/ click "Create payment" in the wizard
6/ Error: 'Customers are required to have a zip, state and country...'

Since payments are not sent to Avatax (only invoices are sent) we
don't want to validate their addresses. The current behavior can also
block payment on a non-Avatax invoice if an Avatax fiscal position is
auto-applied on the payment, which is confusing for users.

This commit ensures to only check non-payments (checked via the
`payment_id` field on `account.move`).

opw-3140130

closes odoo/enterprise#36009

X-original-commit: d30a6d9f665e9b736ed565a90db69304a14b1a84
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: jorenvo <jov@odoo.com>