[FIX] knowledge: directly adding name to article link button
This commit adds the display_name of an article directly as a Text Node
inside of the article link, instead of waiting for the behavior to add it
afterwards.

This will fix any issue linked to HTML field being used inside of an iframe
that needs to access the `/article` command, as seen in the mass_mailing form
view.

The primary issue is, when inside of an iframe like in a mass_mailing body,
the attribute `data-behavior-props` isn't parsed meaning that the link block is
empty in those instances.

To counter this behavior, instead of waiting for the field to be parsed, we
directly add the correct value inside of the block with a text node containing
the value of the `display_name` for the selected article.

task-3116006

closes odoo/enterprise#36644

X-original-commit: 603e4245633b28b845af60b46e4128fe359425f3
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>