[FIX] l10n_de_pos_cert: limit length of buyer with escaped characters
In German localization for Fiskaly, length of the field buyer should
not be more than 50 characters. It has previously been forced in the
template, however if this field has special characters that are
escaped, some extra characters are added for formatting; sometimes
causing it to go over the limit and give an error.

This commit fixes the issue by slicing the escaped string to make
sure it is not more than 50 characters in length.

opw-3125592

closes odoo/enterprise#36077

X-original-commit: 5d47f03217949c832a9a77631d34a8ce7447b9ba
Signed-off-by: Trinh Jacky (trj) <trj@odoo.com>