[IMP] social_*: improve dark mode readability
Purpose:
--------
This commit aims at improving the readability in social modules when the
dark mode is activated.
Therefore, some "raw" color values have been removed, or replaced by
variables that are overridden when the dark mode is set.

Now, when the dark mode is activated, the author and command links in the
comment modals have (more or less) the same colors as on the dark mode
version of the related websites (instead of using the regular ones for
both modes).
The colors of the light mode have also been updated to match the current
designs of these websites in light mode.

Task-3137791

closes odoo/enterprise#35962

Related: odoo/odoo#110736
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>