[FIX] (website_)knowledge: place article in the correct section on frontend
Before this commit, a published shared article was placed into "Public
Articles". This is not correct as they are not public.

If the article was not shared with you specifically, the article was still
available in the public article section. In this case, articles should behave
like published private article. They are available if you have the direct link
but they are not displayed in the side panel.

If the article is shared with you specifically, the article should be in
"Shared with you" section, instead of "Public Articles" section.

After this commit, the articles are listed as the following

  * 'Public Article': Published workspace articles
  * 'Shared with you':
	- non-Published Workspace article you are member of (r/w)
	- Shared article you are member of (r/w)

Thus, published shared articles (not shared with you) are not listed the
same way published private articles are still not listed. Those articles
are accessed using a link or a search but the user is not a current active
member of the article.

Task-3140578

X-original-commit: 5daaeeaf111854ebfecb78e1780d40fc09e309d6
Part-of: odoo/enterprise#36846