[FIX] knowledge: properly evaluate show_sidebar when needed.
The show sidebar condition was incorrectly written. If the force flag
'hide_sidebar' is set to true, there is no need to check if there are
some articles to display in the sidebar. And in the default case when
the flag is set to false, the or condition of show_sidebar was never
evaluated.

This commit ensure that the check for available articles is done when
required.

Task-3140578

Part-of: odoo/enterprise#35967