[FIX] knowledge: always display hidden buttons on touch screens
In Knowledge, some buttons are hidden by default and becomes visible on
mouse hover. When the user uses a touch screen, the user will not be able
to see those buttons and interact with them.

This commit will add a new media query to ensure that the css rules hiding
the buttons become active only when the user has the hover capability.
The touch screen users will then be able to see and interact with all
buttons of Knowledge.

task-2867059

Part-of: odoo/enterprise#37029