[IMP] mrp_{workorder_}plm: create ECO from MO
From a Manufacturing Order, the user can create an Engineering Change
Order through the form view action "Create ECO".
This action will create a new ECO and links it with the MO.
Once the revision is started, the revision's BoM will be created
according the MO's BoM, including the changes done in the MO.

The MO and the ECO are linked thanks to two new field:
- `production_id` for `mrp.eco`: it's the MO who created the ECO;
- `eco_ids` for `mrp.production`: it's all the ECO created by this MO.
A stat button is added in both `mrp.eco` and `mrp.production` form view.

task-3089368

Part-of: odoo/enterprise#35969