[FIX] *_intrastat : intrastat not initialized for sale invoice, backport
Reproduction:
1. Install Accounting, account_intrastat, sale_intrastat
2. Go to Accounting, create a new customer invoice
3. Choose a customer from the US, go to the Other info tab, the
Intrastat Country is unset, and this is correct
4. Edit the shipping address to one in the EU, the Intrastat Country is
still not set, which is incorrect

Reproduction 2:
1. Install Accounting, Sale, account_intrastat, sale_intrastat
2. In a Belgium company, go to Sale, make a sale order
3. Choose a customer from the US, edit the shipping address to one in
the EU but not Belgium
4. Confirm the order, validate the delivery and create the invoice
5. The Intrastat country on the invoice is empty

Reason: the recomputing method to reset the intrastat country was
removed for Odoo 16

Fix: This is a back port of odoo/enterprise#34622 (https://github.com/odoo/enterprise/commit/d96930c706b5e13ec94a4759d2ba0f00b883b9ba)
and added the initialization correction for the following versions

Orininal Fix message for previous ports:
Change the condition based on delivery address. This is not an issue in
Odoo 16 because we have a recompute all compute field step. Nevertheless
this initialization should fw to master for correctness. However, simply
fixing the initialization will conflict with the compute function's
logic. Thus we backport the change in 16 to 14 to have a clean and more
compact compute logic. The computing logic is re-writtenfor easier
understanding in sale_instrastat. Since the partner_shipping_id is not
defined in account in 14, an override of the compute method is used to
add the dependency

opw-3051881

closes odoo/enterprise#38703

X-original-commit: 4f3d5798a42680c7465544fd0bd9f74ee135a4f8
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Liu Jinjiu (jili) <jili@odoo.com>