[FIX] l10n_mx_edi_40: fix CFDI validation for not-mexican customers
When creating an invoice with a customer which country isn't Mexico, the
CFDI validation fails with error : "La clave del campo RegimenFiscalR
debe corresponder con el tipo de persona (física o moral).". Indeed, the
documentation states that when the RFC is 'XEXX010101000', the fiscal
regime should be '616'. In 16.0, a default value (601) for the fiscal
regime has been added (see
https://github.com/odoo/enterprise/commit/8e65041f99f4e687887bd57d59860b76f138052f).

As a consequence, since the fiscal regime is not falsy, the
`_l10n_mx_edi_get_40_values` doesn't reassign the fiscal regime as 616
when the country is different from Mexico, and the CFDI validation
fails. To fix the issue, we allow to reassign the fiscal regime even if
it is already set.

closes odoo/enterprise#35977

Signed-off-by: Josse Colpaert <jco@odoo.com>