[FIX] web_enterprise: view switcher display on legacy control panel
This commit fixes the view switcher display for views using the legacy
control_panel (e.g. the activity view). To fix this behavior, we use the
`usePosition` hook like in the Dropdown component.

Note: As this bug only occurs in legacy views and they won't be present
in next release, there is no test (also it's an edge case and not a
blocking bug).

Steps to reproduce:

1. Open Project on small screen device like a smartphone
2. Open a Project like 'Office Design'
3. Switch to the Activity view and try to switch the view again
   => a part of the dropdown menu is displayed outside the screen
   => bug

closes odoo/enterprise#36056

Signed-off-by: Pierre Paridans (app) <app@odoo.com>