[FIX] website_knowledge: load workspace published articles on home for share users
Before this commit, when no article was shared with the user, the home page
redirected to a 'No Article found' page, even if some workspace article were
published, and thus available for the user. The user needed to type the exact
link of one of the published workspace article to land on the article and see
in the left panel all the published workspace articles available.

After this commit, if there is at least one published workspace article, the
share user will land on the first article found.

Task-3140578

Part-of: odoo/enterprise#35967