[FIX] sale_amazon_spapi: consider if the taxes are included
Before this commit, when synchronizing an order with a discount (either a Promotion Discount or a
Shipping Discount), the Amazon connector would take the value Amazon sent without considering if
the amount was tax included or not, and pass it as is without knowing if it should be tax included
or not, which could lead to some sales order lines having a wrong discount percentage.

The Amazon connector will now take this info from Amazon and recompute it in order to match
the expected value.

task-2700405
opw-3116517

closes odoo/enterprise#37018

X-original-commit: 7d15a296f28264df83b47ea7e89fc1bfebac4ed7
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>
Signed-off-by: Morgane Demesmaeker <edm@odoo.com>