[IMP] l10n_ar_reports: form view for account.ar.vat.line model
before this commit, there is no form view defined for the model account.ar.vat.line and thus on opening the form view by clicking on the pivot view cells, it is returning an odoo generated view, which is not looking great.

after this commit, as form view is defined for the model, clean form view will be shown to the end user.

closes odoo/enterprise#36085

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>