[REF] various: apply main attachment mixin on models that use the feature
Main attachment is quite costly and as it is implemented in mail.thread, is
present in most model. In order to improve performance, we remove it here from
mail.thread and add it, through a new mixin, only on model that actually uses
that features.

In various module, we have added the new mixin in the inheritance chain of the
models that use that feature.

Task-2648976

Part-of: odoo/enterprise#36164