[FIX] web_enterprise: `ui-autocomplete` border
Prior to this commit, when the `ui-autocomplete` dropdown was longer
than the input it came from, or if the dropdown appeared above the input
(when there was not enough space below it), there was no top border and
this created a graphical issue.

After this commit, the dropdown always has a top border and fixes this
issue.

task-3140983

closes odoo/enterprise#39494

X-original-commit: 0879403209e4a9810c42710e695a41d4f3f16e73
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>