[FIX] stock_barcode: scan barcode shared by consu product and lot
A product barcode and a lot name can have the same value. In such
case, it will lead to issues when trying to scan it

To reproduce the issue:
1. Create two products P01, P02
   - P01:
     - Barcode: 123
   - P02:
     - Tracking: By lots
2. Create a lot L for P02:
   - Name: 123
3. Operation Type > Delivery:
   - Can create lot: False
   - Use existing: True
4. Barcodes > Operations > Delivery, New
5. Scan 123
   - A line appears for P01 and a warning is displayed as there is a
     confusion between P01 and L
6. Click somewhere (to trigger a save)

Error: a validation error is raised because of the incompatibility
between L and P01. The user is stuck.

As said above, clicking anywhere will trigger a save of the SML, but
with P01 as product and L as lot. It will trigger a constraint:
https://github.com/odoo/odoo/blob/c9fdd4f2ac2d8afb18cb0dd126bb0b88c54f374e/addons/stock/models/stock_move_line.py#L109-L117

OPW-3100252

closes odoo/enterprise#36755

X-original-commit: d1c93421d4f26d614e43b6fa15bad9ba06dfc073
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>