[FIX] account_reports: groupby expansion: properly apply default values to expressions depending on their figure type
Before, when a report contained multiple expressions on a line using a groupby, if multiple figure types were used on these expressions, only the figure type of the last of these expression was used as a default value for grouped lines (this default value is needed for example when grouping by partner, if partner A is obtained from expression 1, but not in the result of expression 2 => expression 2 displays the default value).

For example, if expression 1 had figure_type 'none' and expression 2 had 'monetary', expression 1's default value would be '0 $' instead of ''.

This was due to bad grouping in the processing of groupby lines expansion. It is now fixed.

closes odoo/enterprise#36111

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>