[FIX] l10n_ar_reports: fix test vat_book with product
The product `product_iva_105_perc` in the test was not created with the copy().
We fixed that, and by doing so, the real taxes now apply.
The amounts of the invoices have changed and so, the export has changed.
We fix the tests to reflect that.

task-3045026

Part-of: odoo/enterprise#38163