[FIX] account_sepa: Never put SEPA SvcLvl in exported XML file for Swiss variant
Before this fix, the SvcLvl node was always added if the file was not detected as generic SEPA. This happened when all the accounts involved were IBAN, and the amounts were all in EUR. When using the Swiss variant, this was a problem, as this value is not supported for the (optional) SvcLvl node. We hence solve the issue by not adding the node for this variant anymore.

OPW 3006565

closes odoo/enterprise#36196

X-original-commit: cf820e6dcb570ee0d36b02e22887af80100beb37
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>