[IMP] account_report: pdf changes
When printing a PDF of any report, the header was incomplete. Indeed the name of the report itself was missing since version 16.0 of Odoo. This PR solves that by adding the report title when the HTML of the pdf is created.
Other bug seen was that since the dark mode, all the separators between lines of the report were missing when printing the report in PDF because wkhtmltopdf doesn't allow the use of VAR in the css file.

closes odoo/enterprise#37524

Task-id: 3135679
X-original-commit: a0368b56f1b9687f9c0f91971e22841b29f9d00a
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>