[FIX] social: prevent opening of form on clicking kanban
before this commit on clicking the kanban of social.media model, it opens the form view of social.media model, and as a form view is not added for the model, it currently opens a odoo defined form view.

after this commit, the kanban will not be clickable as in the previous versions.

introduced by: https://github.com/odoo/enterprise/commit/3d5450231857fa8a660b33f60112e888ed27bf52

closes odoo/enterprise#36165

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>