[FIX] account_followup: email not sent on follow-up
Currently, when sending a follow-up email, the email itself is not actually sent. On top of that, in the chatter, it appears like the email was sent by the customer instead of to the customer.

To fix this, we just need to remove the `email_from` parameter, as it was set to the customer's email.

closes odoo/enterprise#36230

Signed-off-by: Quentin De Paoli <qdp@odoo.com>