[FIX] l10n_es_reports: traceback in 349 trimester report
When generating a BOE 349 Spanish tax report with the "Trimester
monthly report" option enabled, and if there are refunds included in the
report period, you will get an empty traceback. The reason this happens
is that when that report option is enabled, it will reduce the `date_to`
of the report filter by 1 month (because that's what that feature does).
A `_get_mod_period_and_year` call later on uses `date_from` and
`date_to` to determine the year and period and the period calculation
will fail and return None because it's not aware of the modification.
Later on when the period is used and subscripted it will give the
traceback.

The solution used here uses the `options` dict to keep the original
`date_to` for `_get_mod_period_and_year` to use. A cleaner solution
could be to simply pass `year` and `period` to
`_mod_349_write_type2_refund_records` in `_boe_export_mod349` so the
`_get_mod_period_and_year` call isn't needed anymore. But this requires
modifying the function signature for
`_mod_349_write_type2_refund_records`.

The traceback being empty is a separate issue, I've seen many different
kinds of errors occur and they all end up showing as blank. To be fixed
in a different PR.

Steps to reproduce
1. Start from a clean 14.0 database
2. Set the default company to have Spain as the country
3. Install Accounting and Contacts
4. In Contacts, create a new company contact called Odoo and its country to Belgium. What matters is that the country is European but not Spain. Set the VAT number to BE0477472701.
5. Switch to the default ES Company that got added after installing Accounting
6. In Accounting, create a customer invoice for the Odoo contact with some random product that has a price. Set the invoice date to somewhere in Q4 2022 (like 15 November 2022). In the AEAT data tab, set the type for mod 349 to "A - Acquisition".
7. Revert the customer invoice as a full refund and with the same date as the invoice.
8. In Accounting, go to Reporting -> Tax Report (Mod 349), and set the date filter to the last trimester (currently Q4 2022). You should see an amount on the line "A. Adquisiciones intracomunitarias sujetas" under "Reembolsos".
9. Press the "Export (BOE)" button, and check "Trimester monthly report". Click "Generate BOE".
10. Empty traceback

opw-3131030

closes odoo/enterprise#36510

X-original-commit: 2d60555b99522c6c9ab007bc75970a29b7cddd6c
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Merel Geens <mege@odoo.com>