[IMP] account_asset_fleet: add vehicle to bank recon
Users might import bank transactions related to vehicles, and would like to manually reconcile these transactions directly to a vehicle expense account.

Currently, it is not possible to associate this entry with a vehicle.
After this PR, users are able to specify the vehicle in the Manual Operations Tab of the bank reconciliation.

In the case where there is a disallowed expense account with car category, the vehicle field is required.

Task-3142842

closes odoo/enterprise#36240

Signed-off-by: Laurent Smet (las) <las@odoo.com>