[FIX] l10n_be_coda: fix import of accented chars
When importing a CODA file containing accented characters, they can show up as other characters.
This has to do with the file encoding of the CODA file.

e.g. the letter `é` in `UTF-8` is represented as `00C3 00A9` (using 2 bytes).
When we first try parsing in `CP1252`, which uses 1 byte per letter, it will look like `Ã` (`00C3`) and `©` (`00A9`).
These are both printable characters and so we don't complain when decoding the file.

In this commit we change the order of encoding trials to first try UTF-8, which is a sensible default.

Also, the `_check_coda` function did not work for `CP1140`, `UTF-16` and `UTF-32` encoded files.
The regex was defined as a string and converted to bytes using the default ASCII encoding. This didn't match the bytes of the previous three encodings.

In this commit we first try to decode the file and then do a string regex check on the decoded string.

task-3147804

closes odoo/enterprise#36400

X-original-commit: 58bf5d14a23a9d1cff17acef64b1ba0408f3e399
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>